fix: Fix SyntaxWarning for regex patterns and torch_dtype deprecation warning

+++ src/model/__init__.py
@@ -103,7 +103,7 @@ def get_model(model_cfg: DictConfig):
         
         model = model_cls.from_pretrained(
             pretrained_model_name_or_path=model_path,
-            torch_dtype=torch_dtype,
+            dtype=torch_dtype,  # Use dtype instead of deprecated torch_dtype
             **model_args_dict_final,
             cache_dir=hf_home,
         )
